+++ juce/source/bass/ui/BassFxRack.cpp
@@ -28,7 +28,7 @@ void BassFxRack::Group::layout(juce::Rectangle<int> r) {
     const float cw = (float)inner.getWidth() / (float)n;
     const int kh = juce::jmin(inner.getHeight(), Knob::svgPx(Knob::Sm) + 13); // dia + label
     for (int i = 0; i < n; ++i)
-        knobs[i]->setBounds((int)std::round(inner.getX() + i * cw), inner.getY(),
+        knobs[i]->setBounds((int)std::round(static_cast<float>(inner.getX()) + static_cast<float>(i) * cw), inner.getY(),
                             (int)std::round(cw), kh);
 }
 
@@ -77,9 +77,10 @@ BassFxRack::BassFxRack(BassUiModel& p) {
 void BassFxRack::resized() {
     auto r = getLocalBounds().reduced(8);        // .bl-fx-panel padding
     const int gap = 10;
-    const float cw = (r.getWidth() - gap * 3) / 4.0f;
+    const float cw = static_cast<float>(r.getWidth() - gap * 3) / 4.0f;
     for (int i = 0; i < groups.size(); ++i)
-        groups[i]->layout({ (int)std::round(r.getX() + i * (cw + gap)), r.getY(),
+        groups[i]->layout({ (int)std::round(static_cast<float>(r.getX())
+                                             + static_cast<float>(i) * (cw + static_cast<float>(gap))), r.getY(),
                             (int)std::round(cw), r.getHeight() });
 }
